chore: wip

+++ dot/parachain/prospective-parachains/fragment_chain_test.go
@@ -280,7 +280,7 @@ func TestEarliestRelayParent(t *testing.T) {
 					Hash:   common.Hash{0x02},
 					Number: 9,
 				}
-				ancestorsMap := btree.NewMap[uint, relayChainBlockInfo](100)
+				ancestorsMap := btree.NewMap[uint32, relayChainBlockInfo](100)
 				ancestorsMap.Set(ancestor.Number, ancestor)
 				return &scope{
 					relayParent:     relayParent,
@@ -305,7 +305,7 @@ func TestEarliestRelayParent(t *testing.T) {
 				return &scope{
 					relayParent:     relayParent,
 					baseConstraints: baseConstraints,
-					ancestors:       btree.NewMap[uint, relayChainBlockInfo](100),
+					ancestors:       btree.NewMap[uint32, relayChainBlockInfo](100),
 				}
 			},
 			expect: relayChainBlockInfo{
@@ -488,8 +488,8 @@ func TestFragmentChainWithFreshScope(t *testing.T) {
 }
 
 func makeConstraints(
-	minRelayParentNumber uint,
-	validWatermarks []uint,
+	minRelayParentNumber uint32,
+	validWatermarks []uint32,
 	requiredParent parachaintypes.HeadData,
 ) *parachaintypes.Constraints {
 	return &parachaintypes.Constraints{
@@ -499,7 +499,7 @@ func makeConstraints(
 		UmpRemaining:          10,
 		UmpRemainingBytes:     1_000,
 		MaxUmpNumPerCandidate: 10,
-		DmpRemainingMessages:  make([]uint, 10),
+		DmpRemainingMessages:  make([]uint32, 10),
 		HrmpInbound: parachaintypes.InboundHrmpLimitations{
 			ValidWatermarks: validWatermarks,
 		},
@@ -585,7 +585,7 @@ func TestScopeRejectsAncestors(t *testing.T) {
 				},
 			},
 			maxDepth: 2,
-			baseConstraints: makeConstraints(8, []uint{8, 9},
+			baseConstraints: makeConstraints(8, []uint32{8, 9},
 				parachaintypes.HeadData{Data: []byte{0x01, 0x02, 0x03}}),
 			pendingAvailability: make([]*pendingAvailability, 0),
 			expectedError:       errUnexpectedAncestor{number: 8, prev: 10},
@@ -604,7 +604,7 @@ func TestScopeRejectsAncestors(t *testing.T) {
 				},
 			},
 			maxDepth:            2,
-			baseConstraints:     makeConstraints(0, []uint{}, parachaintypes.HeadData{Data: []byte{1, 2, 3}}),
+			baseConstraints:     makeConstraints(0, []uint32{}, parachaintypes.HeadData{Data: []byte{1, 2, 3}}),
 			pendingAvailability: make([]*pendingAvailability, 0),
 			expectedError:       errUnexpectedAncestor{number: 99999, prev: 0},
 		},
@@ -632,7 +632,7 @@ func TestScopeRejectsAncestors(t *testing.T) {
 				},
 			},
 			maxDepth:            2,
-			baseConstraints:     makeConstraints(0, []uint{2}, parachaintypes.HeadData{Data: []byte{1, 2, 3}}),
+			baseConstraints:     makeConstraints(0, []uint32{2}, parachaintypes.HeadData{Data: []byte{1, 2, 3}}),
 			pendingAvailability: make([]*pendingAvailability, 0),
 			expectedError:       errUnexpectedAncestor{number: 2, prev: 4},
 		},
@@ -679,7 +679,7 @@ func TestScopeOnlyTakesAncestorsUpToMin(t *testing.T) {
 	}
 
 	maxDepth := uint(2)
-	baseConstraints := makeConstraints(3, []uint{2}, parachaintypes.HeadData{Data: []byte{1, 2, 3}})
+	baseConstraints := makeConstraints(3, []uint32{2}, parachaintypes.HeadData{Data: []byte{1, 2, 3}})
 	pendingAvailability := make([]*pendingAvailability, 0)
 
 	scope, err := newScopeWithAncestors(relayParent, baseConstraints, pendingAvailability, maxDepth, ancestors)
@@ -943,7 +943,7 @@ func TestCandidateStorageMethods(t *testing.T) {
 }
 
 func TestInitAndPopulateFromEmpty(t *testing.T) {
-	baseConstraints := makeConstraints(0, []uint{0}, parachaintypes.HeadData{Data: []byte{0x0a}})
+	baseConstraints := makeConstraints(0, []uint32{0}, parachaintypes.HeadData{Data: []byte{0x0a}})
 
 	scope, err := newScopeWithAncestors(
 		relayChainBlockInfo{
@@ -1005,7 +1005,7 @@ func TestPopulateAndCheckPotential(t *testing.T) {
 	}
 
 	firstParachainHead := parachaintypes.HeadData{Data: []byte{0x0a}}
-	baseConstraints := makeConstraints(0, []uint{0}, firstParachainHead)
+	baseConstraints := makeConstraints(0, []uint32{0}, firstParachainHead)
 
 	// helper function to hash the candidate and add its entry
 	// into the candidate storage
@@ -1071,10 +1071,10 @@ func TestPopulateAndCheckPotential(t *testing.T) {
 			// define a constraint that requires a parent head data
 			// that is different from candidate A parent head
 			*makeConstraints(relayParentAInfo.Number,
-				[]uint{relayParentAInfo.Number}, parachaintypes.HeadData{Data: []byte{0x0e}}),
+				[]uint32{relayParentAInfo.Number}, parachaintypes.HeadData{Data: []byte{0x0e}}),
 
 			// the min relay parent for candidate A is wrong
-			*makeConstraints(relayParentBInfo.Number, []uint{0}, firstParachainHead),
+			*makeConstraints(relayParentBInfo.Number, []uint32{0}, firstParachainHead),
 		}
 
 		for _, wrongConstraint := range wrongConstraints {
@@ -1834,7 +1834,7 @@ func TestPopulateAndCheckPotential(t *testing.T) {
 			require.ErrorIs(t, chain.canAddCandidateAsPotential(candidateAEntry), errCandidateAlreadyKnown)
 
 			// simulate the fact that candidate A, B and C have been included
-			baseConstraints := makeConstraints(0, []uint{0}, parachaintypes.HeadData{Data: []byte{0x0d}})
+			baseConstraints := makeConstraints(0, []uint32{0}, parachaintypes.HeadData{Data: []byte{0x0d}})
 			scope, err = newScopeWithAncestors(*relayParentCInfo, baseConstraints, nil, 2, ancestors)
 			require.NoError(t, err)
 
@@ -1923,7 +1923,7 @@ func TestFindAncestorPathAndFindBackableChainEmptyBestChain(t *testing.T) {
 	maxDepth := uint(10)
 
 	// Empty chain
-	baseConstraints := makeConstraints(0, []uint{0}, requiredParent)
+	baseConstraints := makeConstraints(0, []uint32{0}, requiredParent)
 
 	relayParentInfo := relayChainBlockInfo{
 		Number:      0,
@@ -2014,12 +2014,12 @@ func TestFindAncestorPathAndFindBackableChain(t *testing.T) {
 	}
 
 	relayParentInfo := relayChainBlockInfo{
-		Number:      uint(relayParentNumber),
+		Number:      relayParentNumber,
 		Hash:        relayParent,
 		StorageRoot: relayParentStorageRoot,
 	}
 
-	baseConstraints := makeConstraints(0, []uint{0}, requiredParent)
+	baseConstraints := makeConstraints(0, []uint32{0}, requiredParent)
 	scope, err := newScopeWithAncestors(
 		relayParentInfo,
 		baseConstraints,
